requirements.txt
pyTelegramBotAPI
google-generativeai
python-dotenv
pillow
requests
duckduckgo-search
pypdf
edge-tts
youtube-transcript-api
flask
